requirements.txt
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
django-prometheus==2.3.1
drf-yasg==1.21.10
parameterized==0.9.0
psycopg2-binary==2.9.10
pydantic==2.11.3
python-decouple==3.8
python-dateutil==2.9.0.post0
